+++ test/next/jdbc/types_test.clj
@@ -2,13 +2,13 @@
 
 (ns next.jdbc.types-test
   "Some tests for the type-assist functions."
-  (:require [clojure.test :refer [deftest is testing]]
+  (:require [clojure.test :refer [deftest is]]
             [next.jdbc.types :refer [as-varchar]]))
 
 (set! *warn-on-reflection* true)
 
 (deftest as-varchar-test
   (let [v (as-varchar "Hello")]
-    (is (= ["Hello"] v))
+    (is (= "Hello" (v)))
     (is (contains? (meta v) 'next.jdbc.prepare/set-parameter))
     (is (fn? (get (meta v) 'next.jdbc.prepare/set-parameter)))))